fix(auth): address PR feedback
